refactors asset package
