+++ packages/assets/src/index.ts
@@ -1,33 +1,33 @@
-import AddImage from "./lib/images/AddImage";
-import AddSong from "./lib/images/AddSong";
+import AddImage from "./lib/icons/AddImage";
+import AddSong from "./lib/icons/AddSong";
 import artistAgreementPreview from "./lib/images/artist-agreement-preview.jpg";
 import bgImage from "./lib/images/bg-img.webp";
-import Charli3Logo from "./lib/images/Charli3Logo";
-import CheckboxIcon from "./lib/images/CheckboxIcon";
-import CheckCircle from "./lib/images/CheckCircle";
-import CheckIcon from "./lib/images/CheckIcon";
-import CloseCircleFill from "./lib/images/CloseCircleFill";
-import CloseIcon from "./lib/images/CloseIcon";
-import DiscordLogo from "./lib/images/DiscordLogo";
-import DropdownAdornment from "./lib/images/DropdownAdornment";
-import FoldersIcon from "./lib/images/FoldersIcon";
-import GlobalFill from "./lib/images/GlobalFill";
-import InfoCircleLine from "./lib/images/InfoCircleLine";
-import InstagramLogo from "./lib/images/InstagramLogo";
-import NEWMLogo from "./lib/images/NEWMLogo";
-import NEWMLogoSmInverse from "./lib/images/NEWMLogoSmInverse";
-import NEWMMarketplaceLogo from "./lib/images/NEWMMarketplaceLogo";
-import Owner from "./lib/images/Owner";
-import PeopleIcon from "./lib/images/PeopleIcon";
-import PlayButton from "./lib/images/PlayButton";
-import SelectedCheckboxIcon from "./lib/images/SelectedCheckboxIcon";
-import SoundWave from "./lib/images/SoundWave";
-import TimeCircleLine from "./lib/images/TimeCircleLine";
-import TwitterLogo from "./lib/images/TwitterLogo";
-import UnselectedCheckboxIcon from "./lib/images/UnselectedCheckboxIcon";
-import UploadIcon from "./lib/images/UploadIcon";
-import VerticalEllipsis from "./lib/images/VerticalEllipsis";
-import WalletIcon from "./lib/images/WalletIcon";
+import Charli3Logo from "./lib/icons/Charli3Logo";
+import CheckboxIcon from "./lib/icons/CheckboxIcon";
+import CheckCircle from "./lib/icons/CheckCircle";
+import CheckIcon from "./lib/icons/CheckIcon";
+import CloseCircleFill from "./lib/icons/CloseCircleFill";
+import CloseIcon from "./lib/icons/CloseIcon";
+import DiscordLogo from "./lib/icons/DiscordLogo";
+import DropdownAdornment from "./lib/icons/DropdownAdornment";
+import FoldersIcon from "./lib/icons/FoldersIcon";
+import GlobalFill from "./lib/icons/GlobalFill";
+import InfoCircleLine from "./lib/icons/InfoCircleLine";
+import InstagramLogo from "./lib/icons/InstagramLogo";
+import NEWMLogo from "./lib/icons/NEWMLogo";
+import NEWMLogoSmInverse from "./lib/icons/NEWMLogoSmInverse";
+import NEWMMarketplaceLogo from "./lib/icons/NEWMMarketplaceLogo";
+import Owner from "./lib/icons/Owner";
+import PeopleIcon from "./lib/icons/PeopleIcon";
+import PlayButton from "./lib/icons/PlayButton";
+import SelectedCheckboxIcon from "./lib/icons/SelectedCheckboxIcon";
+import SoundWave from "./lib/icons/SoundWave";
+import TimeCircleLine from "./lib/icons/TimeCircleLine";
+import TwitterLogo from "./lib/icons/TwitterLogo";
+import UnselectedCheckboxIcon from "./lib/icons/UnselectedCheckboxIcon";
+import UploadIcon from "./lib/icons/UploadIcon";
+import VerticalEllipsis from "./lib/icons/VerticalEllipsis";
+import WalletIcon from "./lib/icons/WalletIcon";
 
 export {
   AddSong,
