+++ packages/assets/src/lib/icons/AddImage.tsx
