+++ packages/assets/src/lib/icons/AddSong.tsx
