+++ packages/assets/src/lib/icons/Charli3Logo.tsx
