+++ packages/assets/src/lib/icons/CheckCircle.tsx
