+++ packages/assets/src/lib/icons/CheckIcon.tsx
