+++ packages/assets/src/lib/icons/CheckboxIcon.tsx
