+++ packages/assets/src/lib/icons/CloseCircleFill.tsx
