+++ packages/assets/src/lib/icons/CloseIcon.tsx
