+++ packages/assets/src/lib/icons/DiscordLogo.tsx
