+++ packages/assets/src/lib/icons/DropdownAdornment.tsx
