+++ packages/assets/src/lib/icons/FoldersIcon.tsx
