+++ packages/assets/src/lib/icons/GlobalFill.tsx
