+++ packages/assets/src/lib/icons/InstagramLogo.tsx
