+++ packages/assets/src/lib/icons/NEWMLogo.tsx
