+++ packages/assets/src/lib/icons/NEWMLogoSmInverse.tsx
