+++ packages/assets/src/lib/icons/PeopleIcon.tsx
