+++ packages/assets/src/lib/icons/PlayButton.tsx
