+++ packages/assets/src/lib/icons/SelectedCheckboxIcon.tsx
