+++ packages/assets/src/lib/icons/SoundWave.tsx
