+++ packages/assets/src/lib/icons/UnselectedCheckboxIcon.tsx
