+++ packages/assets/src/lib/icons/UploadIcon.tsx
