+++ packages/assets/src/lib/icons/VerticalEllipsis.tsx
