+++ packages/assets/src/lib/icons/WalletIcon.tsx
