--- packages/assets/src/lib/images/NEWMMarketplaceLogo.tsx
@@ -1,48 +1,0 @@
-const NEWMMarketplaceLogo = () => {
-  return (
-    <svg
-      fill="none"
-      height="60"
-      viewBox="0 0 168 60"
-      width="168"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clipPath="url(#clip0_2865_3308)">
-        <path
-          d="M116.933 32.683v4.68H99.438V9.279h17.495v4.681h-12.236v6.87h9.611v4.43h-9.611v7.423l12.236.001zM161.938 9.278l-9.168 15.646-6.185-10.493.003 9.823 4.466 7.506h3.446l7.871-13.398v19.002h5.258V9.278h-5.691zM80.118 9.278v28.086h5.259L85.375 19.4l10.521 17.964v-9.849L85.377 9.278h-5.26zM137.059 14.882h-3.434l-7.89 13.678V9.278h-5.259l.007 28.086h5.691l9.167-15.646 9.168 15.646v-9.565l-7.45-12.917z"
-          fill="#fff"
-        ></path>
-        <path
-          d="M8.78 51.209c-2.75-2.743-4.936-6-6.435-9.59-6.402-15.273.757-32.86 15.996-39.276a29.855 29.855 0 0123.045-.063 29.952 29.952 0 019.768 6.558 30.09 30.09 0 016.504 9.826c6.238 15.349-1.119 32.86-16.433 39.118a29.78 29.78 0 01-17.211 1.634A29.856 29.856 0 018.78 51.209zm.43-.424a27.915 27.915 0 009.172 6.11 27.825 27.825 0 0010.81 2.078c15.41-.164 27.77-12.818 27.606-28.256a27.983 27.983 0 00-8.177-19.481A27.757 27.757 0 0029.193 3.09C13.796 2.932 1.19 15.313 1.034 30.737a28.111 28.111 0 002.068 10.848 28.36 28.36 0 006.108 9.2z"
-          fill="url(#paint0_linear_2865_3308)"
-        ></path>
-        <path
-          d="M80.195 41h2.086l2.203 5.375h.094L86.781 41h2.086v8h-1.64v-5.207h-.067l-2.07 5.168h-1.117l-2.07-5.188h-.067V49h-1.64v-8zm11.885 8h-1.813l2.762-8h2.18l2.757 8h-1.812l-2.004-6.172h-.063L92.08 49zm-.114-3.145h4.282v1.32h-4.282v-1.32zM99.37 49v-8h3.157c.604 0 1.119.108 1.546.324.43.214.757.517.981.91.226.391.34.85.34 1.38 0 .53-.115.988-.344 1.37-.229.38-.561.672-.996.875-.432.203-.956.305-1.57.305h-2.114v-1.36h1.84c.323 0 .591-.044.805-.132a.957.957 0 00.476-.399c.107-.177.161-.397.161-.66 0-.265-.054-.49-.161-.672a.981.981 0 00-.48-.414c-.214-.096-.483-.144-.809-.144h-1.14V49h-1.692zm4.321-3.64l1.988 3.64h-1.867l-1.946-3.64h1.825zm3.338 3.64v-8h1.691v3.527h.106L111.704 41h2.027l-2.969 3.582L113.766 49h-2.023l-2.192-3.29-.832 1.017V49h-1.691zm8.024 0v-8h5.391v1.395h-3.7V44.3h3.422v1.394h-3.422v1.91h3.715V49h-5.406zm6.871-6.605V41h6.57v1.395h-2.449V49h-1.672v-6.605h-2.449zM130.015 49v-8h3.156c.607 0 1.124.116 1.551.348.427.229.753.548.977.957.226.406.339.875.339 1.406 0 .531-.114 1-.343 1.406a2.375 2.375 0 01-.996.95c-.433.226-.956.34-1.571.34h-2.011V45.05h1.738c.325 0 .594-.056.805-.168.213-.115.372-.272.476-.473.107-.203.16-.436.16-.7a1.44 1.44 0 00-.16-.694 1.07 1.07 0 00-.476-.465c-.214-.112-.485-.168-.813-.168h-1.141V49h-1.691zm7.562 0v-8h1.691v6.605h3.43V49h-5.121zm8.264 0h-1.812l2.762-8h2.179l2.758 8h-1.812l-2.004-6.172h-.063L145.841 49zm-.113-3.145h4.281v1.32h-4.281v-1.32zm14.086-2.054h-1.711a1.676 1.676 0 00-.192-.59 1.496 1.496 0 00-.371-.445 1.63 1.63 0 00-.523-.282 2.008 2.008 0 00-.637-.097c-.414 0-.775.103-1.082.308-.307.203-.546.5-.715.89-.169.389-.254.86-.254 1.415 0 .57.085 1.05.254 1.438.172.387.412.68.719.878.307.198.663.297 1.066.297.227 0 .436-.03.629-.09.195-.06.369-.147.52-.261a1.51 1.51 0 00.375-.426c.101-.167.172-.357.211-.57l1.711.007a3.181 3.181 0 01-1.04 1.973 3.32 3.32 0 01-1.054.633c-.406.154-.866.23-1.379.23-.714 0-1.352-.161-1.914-.484a3.434 3.434 0 01-1.328-1.402c-.323-.612-.485-1.353-.485-2.223 0-.872.164-1.615.492-2.227.329-.612.774-1.078 1.336-1.398a3.748 3.748 0 011.899-.484c.463 0 .893.065 1.289.195.398.13.751.32 1.059.57.307.248.557.551.75.91.195.36.32.771.375 1.235zM161.466 49v-8h5.391v1.395h-3.699V44.3h3.422v1.394h-3.422v1.91h3.715V49h-5.407z"
-          fill="#fff"
-        ></path>
-      </g>
-      <defs>
-        <linearGradient
-          gradientUnits="userSpaceOnUse"
-          id="paint0_linear_2865_3308"
-          x1="8.262"
-          x2="51.518"
-          y1="51.216"
-          y2="8.624"
-        >
-          <stop stopColor="#FFC33C"></stop>
-          <stop offset="0.198" stopColor="#FF6E32"></stop>
-          <stop offset="0.396" stopColor="#F53C69"></stop>
-          <stop offset="0.594" stopColor="#C341F0"></stop>
-          <stop offset="0.792" stopColor="#5091EB"></stop>
-          <stop offset="1" stopColor="#41BE91"></stop>
-        </linearGradient>
-        <clipPath id="clip0_2865_3308">
-          <path d="M0 0H167.629V60H0z" fill="#fff"></path>
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export default NEWMMarketplaceLogo;
